fix: remove workspace [SP-410]
