+++ src/pages/Teams/Teams.svelte
@@ -1,23 +1,136 @@
 <script lang="ts">
-  // import TeamsList from "$lib/components/teams/teams-list/TeamsList.svelte";
-  // import TeamsContent from "$lib/components/teams/teams-content/TeamsContent.svelte";
-  import { Motion } from "svelte-motion";
-  import UnderDevelopment from "../UnderDevelopment/UnderDevelopment.svelte";
+  import type {
+    TeamRepositoryMethods,
+    TeamServiceMethods,
+    WorkspaceMethods,
+  } from "$lib/utils/interfaces/workspace.interface";
+
+  import WorkspaceContent from "../../lib/components/workspace/WorkspaceContent.svelte";
+  import WorkspaceList from "../../lib/components/workspace/workspace-list/WorkspaceList.svelte";
   import { scaleMotionProps } from "$lib/utils/animations";
+  import { Motion } from "svelte-motion";
+  import { user } from "$lib/store/auth.store";
+  import { onDestroy } from "svelte";
+  import type { Observable } from "rxjs";
+  import type { TeamDocument } from "$lib/database/app.database";
+  import {
+    openedTeam,
+    setCurrentTeam,
+    setOpenedTeam,
+  } from "$lib/store/team.store";
+  import { setCurrentWorkspace } from "$lib/store/workspace.store";
+  import type { CurrentTeam } from "$lib/utils/interfaces";
+  import { TeamViewModel } from "./team.viewModel";
+  export let data: any,
+    handleWorkspaceSwitch: any,
+    handleWorkspaceTab: any,
+    activeSideBarTabMethods: any,
+    collectionsMethods: any;
+  let allTeams = [],
+    currOpenedTeam: CurrentTeam;
+  const _viewModel = new TeamViewModel();
+  const teams: Observable<TeamDocument[]> = _viewModel.teams;
+  const activeTeam: Observable<TeamDocument> = _viewModel.activeTeam;
+  const workspaceMethods: WorkspaceMethods = {
+    handleCreateTab: _viewModel.handleCreateTab,
+  };
+  const teamRepositoryMethods: TeamRepositoryMethods = {
+    modifyTeam: _viewModel.modifyTeam,
+  };
+  const teamServiceMethods: TeamServiceMethods = {
+    inviteMembersAtTeam: _viewModel.inviteMembersAtTeam,
+    removeMembersAtTeam: _viewModel.removeMembersAtTeam,
+    promoteToAdminAtTeam: _viewModel.promoteToAdminAtTeam,
+    demoteToMemberAtTeam: _viewModel.demoteToMemberAtTeam,
+  };
+
+  const userSubscribe = user.subscribe(async (value) => {
+    if (value) await _viewModel.refreshTeams(value._id);
+  });
+  const openedTeamSubscribe = openedTeam.subscribe((value) => {
+    if (value) currOpenedTeam = value;
+  });
+  const tabList = _viewModel.tabs;
+  const collectionList = _viewModel.collection;
+  let activeTeamRxDoc: TeamDocument;
+  const teamSubscribe = teams.subscribe((value: TeamDocument[]) => {
+    if (value && value.length > 0) {
+      const teamArr = value.map((teamDocument: TeamDocument) => {
+        const teamObj = _viewModel.getTeamDocument(teamDocument);
+        return teamObj;
+      });
+      allTeams = teamArr;
+      if (!activeTeamRxDoc) {
+        _viewModel.activateTeamWorkspace(
+          value[0].get("teamId"),
+          value[0].get("workspaces")[0].workspaceId,
+        );
+        setCurrentTeam(value[0].get("teamId"), value[0].get("name"));
+        setCurrentWorkspace(
+          value[0].get("workspaces")[0].workspaceId,
+          value[0].get("workspaces")[0].name,
+        );
+      }
+      setOpenedTeam(
+        currOpenedTeam.id ? currOpenedTeam.id : value[0].get("teamId"),
+        currOpenedTeam.name ? currOpenedTeam.name : value[0].get("name"),
+      );
+    }
+  });
+
+  const activeTeamSubscribe = activeTeam.subscribe((value: TeamDocument) => {
+    if (value) {
+      activeTeamRxDoc = value;
+    }
+  });
+
+  onDestroy(() => {
+    userSubscribe();
+    openedTeamSubscribe();
+    teamSubscribe.unsubscribe();
+    activeTeamSubscribe.unsubscribe();
+  });
 </script>
 
 <Motion {...scaleMotionProps} let:motion>
-  <div class="d-flex" use:motion>
-    <!-- <TeamsList /> -->
-    <!-- <TeamsContent /> -->
-    <UnderDevelopment
-      heading="Teams"
-      description="Elevating team collaboration seamlessly by leveraging features that streamline onboarding, foster team discovery, and facilitate shared workspace for enhanced cooperation on API projects."
-      subHeading="Teams is set to provide you the power of:"
-      subSubHeadingOne="Unified Collaboration Interface"
-      subSubDescOne="Enhanced teamwork efficiency with centralized interface, fostering seamless collaboration on API projects through discussions, comments, and shared workspaces for request, versions, schemas and scripts. "
-      subSubHeadingTwo="Multi-team Collaboration Flexibility"
-      subSubDescTwo="Seamlessly create, join, and switch between teams, empowering users to manage diverse projects effortlessly across various structures. "
+  <div class="workspace bg-backgroundColor" use:motion>
+    <WorkspaceList
+      teams={allTeams}
+      {data}
+      tabList={$tabList}
+      collectionList={$collectionList}
+      {handleWorkspaceSwitch}
+      {handleWorkspaceTab}
+      {activeSideBarTabMethods}
+      {collectionsMethods}
+    />
+    <WorkspaceContent
+      {handleWorkspaceSwitch}
+      {handleWorkspaceTab}
+      {data}
+      {workspaceMethods}
+      activeTeam={$activeTeam}
+      {activeSideBarTabMethods}
+      {teamServiceMethods}
+      {teamRepositoryMethods}
     />
   </div>
 </Motion>
+
+<style>
+  .workspace {
+    font-size: 12px;
+    top: 44px;
+    left: 352px;
+    width: calc(100% - 352px);
+    position: fixed;
+    height: calc(100% - 44px);
+    overflow: auto;
+  }
+  .workspace::-webkit-scrollbar {
+    width: 2px;
+  }
+  .workspace::-webkit-scrollbar-thumb {
+    background: #888;
+  }
+</style>
